requirements.txt
openai
google-auth
gspread